+++ src/pages/HomePage/HomePage.styles.js
@@ -1,8 +1,6 @@
 import styled from 'styled-components';
 
-// 전체 게시물 리스트
 export const PostList = styled.ul`
-  /* background-color: green; */
   list-style-type: none;
   padding: 20px;
   width: 1320px;
@@ -13,7 +11,6 @@ export const PostList = styled.ul`
   flex-wrap: wrap;
 `;
 
-// 각각의 게시물
 export const PostItem = styled.li`
   cursor: pointer;
 
@@ -29,7 +26,6 @@ export const PostItem = styled.li`
   }
 `;
 
-// 게시물 타이틀
 export const PostTitle = styled.h2`
   color: black;
   font-size: 24px;
@@ -42,7 +38,6 @@ export const PostTitle = styled.h2`
   text-overflow: ellipsis;
 `;
 
-// 게시물 내용
 export const PostContent = styled.p`
   color: black;
   white-space: pre-wrap;
@@ -56,7 +51,6 @@ export const PostContent = styled.p`
   text-overflow: ellipsis;
 `;
 
-// 게시글 수정 버튼
 export const Button = styled.button`
   margin: 10px 0px 0px 25px;
   padding: 8px 12px;
@@ -70,21 +64,12 @@ export const Button = styled.button`
   }
 `;
 
-// 프로필 이미지
-export const ProfileImage = styled.img`
-  border-radius: 50%;
-  cursor: pointer;
-  width: 45px;
-  height: 45px;
-`;
-
 export const Section = styled.section`
   display: flex;
   gap: 10px;
   align-items: center;
 `;
 
-// 게시물 이미지
 export const PostImage = styled.img`
   width: 250px;
   height: 150px;
